+++ driver/printFeatureVector.h
@@ -2,5 +2,5 @@
 #include <vector>
 #include <iostream>
 
-void printFeatureVector(std::vector<float> featureV);
-void printFeatureVectorPointer(std::vector<float>* featureV);
+void printFeatureVector(std::vector<double> featureV);
+void printFeatureVectorPointer(std::vector<double>* featureV);
